requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.4.13=h06a4308_1
certifi=2020.12.5=py38h06a4308_0
chardet=4.0.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
h5py=3.2.1=pypi_0
idna=2.10=pypi_0
ignite=0.4.4=py_0
intel-openmp=2021.2.0=h06a4308_610
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.0=h27cfd23_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py38h27cfd23_1
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.1=py38ha9443f7_2
ncurses=6.2=he6710b0_1
nettle=3.7.2=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numpy=1.20.1=py38h93e21f0_0
numpy-base=1.20.1=py38h7d8b39e_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
ordered-set=4.0.2=pypi_0
pillow=8.2.0=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
python=3.8.8=hdb3f193_5
pytorch=1.8.1=py3.8_cuda10.2_cudnn7.6.5_0
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=py38h06a4308_0
sqlite=3.35.4=hdfb4753_0
tk=8.6.10=hbc83047_0
torchmeta=1.7.0=pypi_0
torchvision=0.9.1=py38_cu102
tqdm=4.60.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
